Parameters: Moving ISA/platform to `__init__` and `infer_cpu` to tools
